2nd commit: started rebuilding landing page; removed all pictures from page and started inputting css and text
