+++ empoweredconvofrontend/src/components/LandingPage/LandingOld.css
@@ -1,126 +1,423 @@
+/* ------------------------------------*\
+RESET
+\*------------------------------------*/
+
+/* Use your own code or past solution for Great Idea Web Page CSS here! */
+
+/* http://meyerweb.com/eric/tools/css/reset/ 
+   v2.0 | 20110126
+   License: none (public domain)
+*/
+
+html,
+body,
+div,
+span,
+applet,
+object,
+iframe,
+h1,
+h2,
+h3,
+h4,
+h5,
+h6,
+p,
+blockquote,
+pre,
+a,
+abbr,
+acronym,
+address,
+big,
+cite,
+code,
+del,
+dfn,
+em,
+img,
+ins,
+kbd,
+q,
+s,
+samp,
+small,
+strike,
+strong,
+sub,
+sup,
+tt,
+var,
+b,
+u,
+i,
+center,
+dl,
+dt,
+dd,
+ol,
+ul,
+li,
+fieldset,
+form,
+label,
+legend,
+table,
+caption,
+tbody,
+tfoot,
+thead,
+tr,
+th,
+td,
+article,
+aside,
+canvas,
+details,
+embed,
+figure,
+figcaption,
+footer,
+header,
+hgroup,
+menu,
+nav,
+output,
+ruby,
+section,
+summary,
+time,
+mark,
+audio,
+video {
+  margin: 0;
+  padding: 0;
+  border: 0;
+  font-size: 100%;
+  font: inherit;
+  vertical-align: baseline;
+}
+
+/* HTML5 display-role reset for older browsers */
+
+article,
+aside,
+details,
+figcaption,
+figure,
+footer,
+header,
+hgroup,
+menu,
+nav,
+section {
+  display: block;
+}
+body {
+  line-height: 1;
+}
+ol,
+ul {
+  list-style: none;
+}
+blockquote,
+q {
+  quotes: none;
+}
+blockquote:before,
+blockquote:after,
+q:before,
+q:after {
+  content: '';
+  content: none;
+}
+table {
+  border-collapse: collapse;
+  border-spacing: 0;
+}
+
+/*------------------------------------*\
+my CSS
+\*------------------------------------*/
+
+* {
+  box-sizing: border-box;
+}
+html {
+  font-size: 62.5%;
+}
+html,
+body {
+  font-family: 'Roboto', Arial, sans-serif;
+}
+
+code {
+  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+    monospace;
+}
+
 .landing {
-    /* border: 1px solid red; */
-    max-width: 1440px;
-    display: block;
-    margin: 0 auto;
-  }
-  
-  .empowered {
+  margin: 0 auto;
+  max-width: 1440px;
+  overflow: auto;
+  /* Weird quirk: When you remove the border from the div the layout changes. You can prevent margins from collapsing by either adding a border or some padding to the top of your block element, or by creating a new block formatting context by setting overflow: auto */
+}
+
+.landingBackground {
+  position: absolute;
+  z-index: -10;
+}
+
+.intro {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  /* Weird quirk: if you don't define a width on an image and it's in a flex-container, it will stretch to 100%. align-self: [flex-start, center...others] will prevent the image from taking 100% of the flex-containers width.  */
+  margin: 150px 0 0 200px;
+}
+
+.intro img {
+  padding: 0 0 25px 0;
+}
+
+.introMobile {
+  display: none;
+}
+
+.btnExplore {
+  margin: 20px 0 10px 650px;
+}
+
+.ctaContainer {
+  display: flex;
+  justify-content: center;
+  padding-right: 50px;
+}
+
+.cta {
+  margin: 200px 0 0 0;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.ctaBackground {
+  position: absolute;
+  z-index: -9;
+  padding: 110px 0 0 0;
+}
+
+.ctaVideoContainer {
+  position: relative;
+  /* border: 1px solid red; */
+}
+
+.ctaVideo {
+  width: 769px;
+}
+
+.ctaVideoMobile {
+  display: none;
+}
+
+.btnPlay {
+  position: absolute;
+  top: 180px;
+  left: 320px;
+}
+
+.ctaText {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-end;
+  margin: 0 40px 0 0;
+}
+
+.ctaTextMobile {
+  display: none;
+}
+
+.ctaText img {
+  padding: 25px 0;
+}
+
+/* What we do */
+
+.whatWeDoContainer {
+  margin: 170px 0 100px 160px;
+}
+
+.numListItem {
+  display: flex;
+  align-items: flex-start;
+  margin: 60px 0 0 0;
+}
+
+.oval {
+  position: relative;
+}
+
+.ovalNum {
+  position: absolute;
+  top: 15px;
+  left: 18px;
+}
+
+.ovalDescription {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  margin: 13px 0 0 30px;
+}
+
+.numTitle {
+  margin: 0 0 20px 0;
+}
+
+.mobileListItems {
+  display: none;
+}
+
+/* Tablet */
+
+@media (max-width: 1000px) {
+  .introContainer {
     display: flex;
+    justify-content: center;
   }
-  
-  .empoweredLeft {
-    width: 66%;
+
+  .introMobile {
+    display: none;
   }
-  
-  .empoweredLeftBoxes {
-    /* border: 1px solid green; */
-    padding: 145px 0 0 200px;
+
+  .intro {
+    margin: 150px 0 0 0;
   }
-  
-  .empoweredRight {
-    height: 600px;
-    background-color: #24b4a5;
-    width: 34%;
-  }
-  
-  .intro h1 {
-    font-size: 5.5rem;
-    font-weight: bold;
-    padding: 10px 0;
-  }
-  
-  .power {
-    color: #24b4a5;
-  }
-  
-  .intro h2 {
-    font-size: 2.5rem;
-    padding: 0 0 15px 0;
-  }
-  
-  .intro h3 {
-    font-size: 2.8rem;
-    font-weight: bold;
-    font-style: italic;
-    padding: 8px 0 12px 0;
-  }
-  
-  .intro h4 {
-    font-size: 1.8rem;
-    padding: 5px 0;
-  }
-  
-  .btn {
-    display: flex;
-    justify-content: flex-end;
-  }
-  
+
   .btnExplore {
-    background-color: #2d2f38;
-    border: none;
-    border-radius: 3px;
-    padding: 10px 50px;
-    margin: 45px 140px 45px 0;
-    font-size: 1.7rem;
-    color: white;
+    margin: 20px 0 10px 450px;
   }
-  
-  .introBottomBox {
+
+  .ctaContainer {
+    padding: 200px 0 0 0;
     display: flex;
-    justify-content: flex-end;
+    flex-direction: column;
   }
-  
-  .introBottomBoxRight {
-    height: 107px;
-    background-color: #8fedc2;
-    width: 50%;
+
+  .cta {
+    margin: 0 0 0 0;
   }
-  
-  /* cta mid section */
-  
-  .cta {    
-    position: relative;
+
+  .ctaText {
+    display: none;
   }
-  
-  .ctaBoxes {
-    /* border: 1px solid green; */
+
+  .ctaTextMobile {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-end;
+    margin-right: 100px;
+  }
+
+  .whatWeDoContainer {
+    margin: 0 auto;
+    padding: 170px 0 40px 0;
+  }
+
+  .whatWeDoImg {
+    display: block;
+    margin-left: auto;
+    margin-right: auto;
+  }
+
+  .numListItem {
+    display: none;
+  }
+
+  .mobileListItems {
     display: flex;
   }
-  
-  .ctaBoxesLeft {
-    width: 66%;
+
+  .mobileListItem {
     display: flex;
-    justify-content: flex-end;
+    flex-direction: column;
+    align-items: center;
+    width: 100%;
   }
-  
-  .ctaBoxesMid {
-    height: 410px;
-    background-color: #8fedc2;
-    width: 50%;
+
+  .mobileListItemTitle {
+    margin: 50px 0 20px 0;
   }
-  
-  .ctaVideo {
+}
+
+@media (max-width: 790px) {
+
+  .landingBackground {
+    display: none;
     position: absolute;
-    top: 80px;
-    left: 200px;
-    background-color: grey;
-    height: 480px;
-    width: 720px;
-    color: white;
-    text-align: center;
-    font-size: 5rem;
-    padding-top: 200px;
+    z-index: -10;
   }
-  
-  .ctaBoxesRight {
-    height: 410px;
+
+  .ctaContainer {
+    background-color: #8FEDC2;
+    padding: 100px 0 100px 0;
+  }
+
+  .ctaBackground {
+    display: none;
+    background-color: #8FEDC2;
+  }
+
+  .ctaTextMobile {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-end;
+    margin-right: 20px;
+  }
+
+  .mobileListItems {
+    flex-direction: column;
     background-color: #24b4a5;
-    width: 34%;
+    padding: 0 0 50px 0;
   }
-  
-  .ctaBoxesBottom {
-    height: 250px;
-    background-color: #24b4a5;
+
+  .introDesktop {
+    display: none;
   }
-  
+
+  .introMobile {
+    display: block;
+  }
+
+  .intro {
+    margin: 100px 0 0 0;
+  }
+
+  .btnExplore {
+    margin: 20px 0 10px 160px;
+  }
+  .ctaVideoContainer {
+    height: 328px;
+    margin: 60px 0;
+  }
+
+  .ctaVideo {
+    display: none;
+  }
+
+  .ctaVideoMobile {
+    display: block;
+    width: 100%;
+  }
+
+  .btnPlay {
+    position: absolute;
+    top: 30%;
+    left: 35%;
+  }
+
+  .whatWeDoContainer {
+    padding: 0 0 40px 0;
+  }
+
+}
